+++ grails-app/controllers/au/org/ala/phyloviz/TreeController.groovy
@@ -275,12 +275,17 @@ class TreeController extends BaseController {
     def deleteTree() {
         if (!params.id) {
             badRequest "id is a required parameter"
-        } else if (!Tree.findById(params.id)) {
-            notFound "No tree found for id ${params.id}"
         } else {
-            treeService.deleteTree(params.id as int)
+            Tree tree = Tree.findById(params.id)
+            if (!tree) {
+                notFound "No tree found for id ${params.id}"
+            } else if (!params.isAdmin && tree.getOwner() != treeService.getCurrentOwner()) {
+                notAuthorised "Only the tree owner or an administrator can delete this tree"
+            } else {
+                treeService.deleteTree(params.id as int)
 
-            redirect controller: "wizard", action: "myTrees"
+                redirect controller: "wizard", action: "myTrees"
+            }
         }
     }
